chore(test): convert most of the tests into integration tests
Also improves some test cases here and there and removes redundant ones.
There's still a lot to do to improve the tests in general.
