+++ Cargo.toml
@@ -49,6 +49,7 @@ version = "0.1.9"
 [dev-dependencies]
 criterion = "0.4"
 assert-json-diff = "2.0.2"
+indoc = "1.0"
 pretty_assertions = "1.3.0"
 textwrap = "0.16.0"
