+++ src/parser/tests.rs
@@ -1,17 +1,5 @@
 use super::*;
-use crate::expr::{
-    BinaryOp, BinaryOperator, Expression, Heredoc, HeredocStripMode, ObjectKey, Operation,
-    TemplateExpr, Traversal, TraversalOperator, Variable,
-};
-use crate::structure::{Attribute, Block, Body};
-use crate::template::{IfDirective, StripMode, Template};
 use pest::*;
-use pretty_assertions::assert_eq;
-
-#[track_caller]
-fn expect_body(input: &str, expected: Body) {
-    assert_eq!(parse(input).unwrap(), expected);
-}
 
 #[test]
 fn parse_identifier() {
@@ -322,29 +310,6 @@ fn parse_cond_in_interpolation() {
 }
 
 #[test]
-fn parse_object_with_variable_expr_key() {
-    let input = "providers = { aws.eu-central-1 = aws.eu-central-1 }";
-
-    let expected = Body::builder()
-        .add_attribute((
-            "providers",
-            Expression::from_iter([(
-                ObjectKey::Expression(Expression::from(Traversal::new(
-                    Variable::unchecked("aws"),
-                    [Identifier::unchecked("eu-central-1")],
-                ))),
-                Expression::from(Traversal::new(
-                    Variable::unchecked("aws"),
-                    [Identifier::unchecked("eu-central-1")],
-                )),
-            )]),
-        ))
-        .build();
-
-    expect_body(input, expected);
-}
-
-#[test]
 fn parse_nested_function_call_with_splat() {
     parses_to! {
         parser: HclParser,
@@ -399,49 +364,6 @@ fn parse_nested_function_call_with_splat() {
 }
 
 #[test]
-fn parse_traversal_with_expression() {
-    let input =
-        "route_table_id = aws_route_table.private[count.index % var.availability_zone_count].id";
-
-    let expected = Body::builder()
-        .add_attribute((
-            "route_table_id",
-            Expression::from(Traversal::new(
-                Variable::unchecked("aws_route_table"),
-                [
-                    TraversalOperator::GetAttr("private".into()),
-                    TraversalOperator::Index(Expression::from(Operation::Binary(BinaryOp::new(
-                        Traversal::new(
-                            Variable::unchecked("count"),
-                            [TraversalOperator::GetAttr("index".into())],
-                        ),
-                        BinaryOperator::Mod,
-                        Traversal::new(
-                            Variable::unchecked("var"),
-                            [TraversalOperator::GetAttr("availability_zone_count".into())],
-                        ),
-                    )))),
-                    TraversalOperator::GetAttr("id".into()),
-                ],
-            )),
-        ))
-        .build();
-
-    expect_body(input, expected);
-}
-
-#[test]
-fn parse_null_in_variable_expr() {
-    let input = "foo = null_foo";
-
-    let expected = Body::builder()
-        .add_attribute(("foo", Variable::unchecked("null_foo")))
-        .build();
-
-    expect_body(input, expected);
-}
-
-#[test]
 fn parse_escaped_slash_in_string() {
     parses_to! {
         parser: HclParser,
@@ -453,153 +375,4 @@ fn parse_escaped_slash_in_string() {
             ])
         ]
     };
-}
-
-#[test]
-fn parse_hcl() {
-    let input = r#"
-        resource "aws_s3_bucket" "mybucket" {
-          bucket        = "mybucket"
-          force_destroy = true
-
-          server_side_encryption_configuration {
-            rule {
-              apply_server_side_encryption_by_default {
-                kms_master_key_id = aws_kms_key.mykey.arn
-                sse_algorithm     = "aws:kms"
-              }
-            }
-          }
-        }"#;
-
-    let expected = Body::builder()
-        .add_block(
-            Block::builder("resource")
-                .add_label("aws_s3_bucket")
-                .add_label("mybucket")
-                .add_attribute(Attribute::new("bucket", "mybucket"))
-                .add_attribute(Attribute::new("force_destroy", true))
-                .add_block(
-                    Block::builder("server_side_encryption_configuration")
-                        .add_block(
-                            Block::builder("rule")
-                                .add_block(
-                                    Block::builder("apply_server_side_encryption_by_default")
-                                        .add_attribute(Attribute::new(
-                                            "kms_master_key_id",
-                                            Traversal::new(
-                                                Variable::unchecked("aws_kms_key"),
-                                                ["mykey", "arn"],
-                                            ),
-                                        ))
-                                        .add_attribute(Attribute::new("sse_algorithm", "aws:kms"))
-                                        .build(),
-                                )
-                                .build(),
-                        )
-                        .build(),
-                )
-                .build(),
-        )
-        .build();
-
-    expect_body(input, expected);
-}
-
-#[test]
-fn unescape_strings() {
-    let input = r#"
-        block "label\\with\\backslashes" {
-          string_attr = "I \u2665 unicode"
-
-          object_attr = {
-            "key\nwith\nnewlines" = true
-          }
-
-          heredoc = <<-EOS
-            heredoc template with \
-            escaped newline and \\backslash is not unescaped yet
-          EOS
-        }
-    "#;
-
-    let expected = Body::builder()
-        .add_block(
-            Block::builder("block")
-                .add_label("label\\with\\backslashes")
-                .add_attribute(("string_attr", "I \u{2665} unicode"))
-                .add_attribute((
-                    "object_attr",
-                    Expression::from_iter([("key\nwith\nnewlines", true)]),
-                ))
-                .add_attribute((
-                    "heredoc",
-                    TemplateExpr::Heredoc(
-                        Heredoc::new(
-                            Identifier::unchecked("EOS"),
-                            "            heredoc template with \\\n            escaped newline and \\\\backslash is not unescaped yet\n"
-                        )
-                        .with_strip_mode(HeredocStripMode::Indent)
-                    )
-                ))
-                .build(),
-        )
-        .build();
-
-    expect_body(input, expected);
-}
-
-#[test]
-fn parse_negative_numbers() {
-    let input = r#"
-        float = -4.2
-        float_exp = -4.2e10
-        signed = -42
-    "#;
-
-    let expected = Body::builder()
-        .add_attribute(("float", -4.2f64))
-        .add_attribute(("float_exp", -4.2e10f64))
-        .add_attribute(("signed", -42))
-        .build();
-
-    expect_body(input, expected);
-}
-
-#[test]
-fn parse_template_expr() {
-    let input = r#"foo = "bar ${baz} %{~ if cond}qux%{ endif ~}""#;
-
-    let expected = Body::builder()
-        .add_attribute((
-            "foo",
-            TemplateExpr::QuotedString("bar ${baz} %{~ if cond}qux%{ endif ~}".into()),
-        ))
-        .build();
-
-    let body = parse(input).unwrap();
-
-    assert_eq!(body, expected);
-
-    match body.attributes().next().unwrap().expr() {
-        Expression::TemplateExpr(expr) => {
-            let template = Template::from_expr(expr).unwrap();
-
-            let expected_template = Template::new()
-                .add_literal("bar ")
-                .add_interpolation(Variable::unchecked("baz"))
-                .add_literal(" ")
-                .add_directive(
-                    IfDirective::new(
-                        Variable::unchecked("cond"),
-                        Template::new().add_literal("qux"),
-                    )
-                    .with_if_strip(StripMode::Start)
-                    .with_endif_strip(StripMode::End),
-                );
-
-            assert_eq!(template, expected_template);
-        }
-        _ => unreachable!(),
-    }
 }
